pyproject.toml
[project]
name = "ccxt-mcp"
version = "0.1.0"
description = "MCP server for CCXT trading functionality"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "ccxt",
    "python-dotenv",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ccxt_mcp"]

[tool.hatch.build.targets.sdist]
packages = ["src/ccxt_mcp"] 
